models/open_model_4/requirements.txt
numerapi===2.15.1
python-dotenv===1.0.0
pandas===2.1.0
lightgbm===4.1.0
pyarrow===13.0.0
psutil===5.9.5
numpy===1.26.0
scipy===1.11.3
optuna===3.3.0
google-cloud-storage===2.11.0